Add AppLockTimerProtocol and AppLockTimer class to implement this protocol
